Cached settings now saved into a readable file
With this change the settings are moved from exe.config file as I wanted better config reading and visibility. With this the users can change anytime the settings in config, but if they made a mistake that will be ignored and in the next exit from application it will generate new one.

- Added option to move files into recycle bin instead of removing
- Added installer wizard to simplify the usage, it is stored in AppData/Roaming directory in order to grant ful read/write access to it.